move lower serialization levels to separate file to unbreak import cycle
